chore: upgrade go-ethereum to v1.14.12 and fix CI
- Remove local replace directive for CI compatibility
- Update types.NewBlock API calls for v1.14
- Fix Memory.GetCopy type from int64 to uint64
- Remove deprecated TerminalTotalDifficultyPassed field
- Add gosec rule exclusions for development patterns

+++ pkg/tracing/tracer.go
@@ -180,7 +180,7 @@ func (t *CallTracer) CaptureState(pc uint64, op vm.OpCode, gas, cost uint64, sco
 
 		data := make([]byte, mSize)
 		if mSize > 0 {
-			copy(data, scope.Memory.GetCopy(int64(mStart), int64(mSize)))
+			copy(data, scope.Memory.GetCopy(mStart, mSize))
 		}
 
 		log := CallLog{
